requirements.txt
# FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Image processing
Pillow==10.1.0

# Machine learning and model dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
diffusers>=0.24.0

# Configuration management
PyYAML==6.0.1

# Logging and monitoring
structlog==23.2.0

# Security and rate limiting
slowapi==0.1.9
redis==5.0.1

# System monitoring
psutil==5.9.6

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development dependencies
black==23.11.0
isort==5.12.0
flake8==6.1.0